+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.98.2", date: "2026-03-31" };
+  return { version: "8.98.3", date: "2026-03-31" };
 }